requirements.txt
# Core dependencies
python-dateutil>=2.8.2
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# Azure SDK
azure-identity>=1.14.0
azure-mgmt-monitor>=6.0.0
azure-storage-blob>=12.19.0

# GCP SDK
google-cloud-logging>=3.8.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Machine Learning (includes IsolationForest for anomaly detection)
scikit-learn>=1.3.0
scipy>=1.11.0

# Alerting
slack-sdk>=3.23.0
sendgrid>=6.10.0

# Database (for storing alerts/events)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Logging
structlog>=23.2.0
colorlog>=6.7.0

# Configuration
jsonschema>=4.19.0

# Utilities
click>=8.1.7
tabulate>=0.9.0
cron-descriptor>=1.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
faker>=19.11.0

# Development
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
isort>=5.12.0

# Monitoring
prometheus-client>=0.18.0
